pyproject.toml
[project]
name = "tweetfavtracker"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "maoki", email = "michihito.aoki@gmail.com" }
]
dependencies = [
    "pandas>=2.2.2",
    "openpyxl>=3.1.4",
    "requests>=2.32.3",
    "numpy>=1.26.4",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pip>=24.0",
    "ruff>=0.4.8",
    "spyder-kernels==2.5.*",
    "jupyterlab>=4.2.2",
    "notebook>=7.2.1",
    "jupyterlab-lsp>=5.1.0",
    "python-lsp-server>=1.11.0",
    "python-lsp-ruff>=2.2.1",
    "jupyterlab-code-formatter>=2.2.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/tweetfavtracker"]
